Better separation of layouts and containers, size request optimization.
Layouts don't request the sizes of the cells anymore: that's the container's
job.  Therefore `requestSize` disappears from layouts and `_requestSize` is
renames `requestSize`.  This also clarifies the code.

It is now possible to request the size of a container without making this
container request the size of its children: parameter `forward_request` of
the method `requestSize`.

No more EXPAND_* constants on Cells, just use the strings. Readibility++.